feat(api): implement basic interfacing apis for routes
